Fix spelling and move decompress to utils
