+++ resolwe_bio/management/commands/generate_diffexpr_cuffdiff.py
@@ -24,7 +24,7 @@ from django.utils import timezone
 from resolwe.flow.models import Data, Storage
 from resolwe.utils import BraceMessage as __
 from resolwe_bio.models import Sample
-from .utils import get_descriptorschema, get_process, get_superuser, generate_sample_desciptor
+from .utils import get_descriptorschema, get_process, get_superuser, generate_sample_descriptor
 
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
@@ -88,7 +88,7 @@ class Command(BaseCommand):
 
             sample = Sample.objects.filter(data=exp)[0]
             sample.presample = False
-            sample.descriptor = generate_sample_desciptor('Hs_')
+            sample.descriptor = generate_sample_descriptor('Hs_')
             sample.save()
 
             with open(os.path.join(self.data_dir, str(exp.id), 'stdout.txt'), 'w') as stdout:
