+++ resolwe_bio/management/commands/utils.py
@@ -1,14 +1,17 @@
 """.. Ignore pydocstyle D400.
 
-=============================
-Uility Functions for Commands
-=============================
+==============================
+Utility Functions for Commands
+==============================
 
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import datetime
+import gzip
+import os
 import random
+import zipfile
 
 from django.contrib.auth import get_user_model
 
@@ -42,7 +45,7 @@ def get_superuser():
     return users.first()
 
 
-def generate_sample_desciptor(organism_name):
+def generate_sample_descriptor(organism_name):
     """Generate sample descriptor."""
     annotator = ['Billie Joe Armstrong', 'Dexter Holland', 'Mark Hoppus']
     if organism_name.startswith('Dd'):
@@ -150,3 +153,47 @@ def generate_reads_descriptor(organism_name, presample=True):
     else:
         sample_descriptor['reads_info'] = reads_info
         return sample_descriptor
+
+
+def decompress(file_name):
+    """Compression-agnostic iterator.
+
+    Iterate over files on the archive and return a tuple of
+    file name, line count and file descriptor.
+
+    Supported file formats are .tab, .gz and .zip.
+
+    """
+    if not os.path.isfile(file_name):
+        raise ValueError("Can not find file '{}'".format(file_name))
+
+    line_count = -1
+    _, ext = os.path.splitext(file_name)
+
+    _open = None
+    if ext == '.tab':
+        _open = open
+    elif ext == '.gz':
+        _open = gzip.open
+
+    if _open:
+        with _open(file_name) as tsv_file:
+            line_count = sum(1 for row in tsv_file)
+
+        with _open(file_name) as tsv_file:
+            yield (os.path.basename(file_name), line_count, tsv_file)
+
+    elif ext == '.zip':
+        with zipfile.ZipFile(file_name) as archive:
+            for entry in archive.infolist():
+                if not entry.filename.endswith('.tab'):
+                    continue
+
+                with archive.open(entry) as tsv_file:
+                    line_count = sum(1 for row in tsv_file)
+
+                with archive.open(entry) as tsv_file:
+                    yield (entry.filename, line_count, tsv_file)
+
+    else:
+        raise ValueError("Unsupported file format")
